+++ Cache/Engine/EngineInterface.php
@@ -1,8 +1,9 @@
 <?php
 /**
  *
- * @package     OpenFlame Web Framework
- * @copyright   (c) 2010 OpenFlameCMS.com
+ * @package     openflame-framework
+ * @subpackage  cache
+ * @copyright   (c) 2010 - 2011 openflame-project.org
  * @license     http://opensource.org/licenses/mit-license.php The MIT License
  * @link        https://github.com/OpenFlame/OpenFlame-Framework
  *
@@ -12,10 +13,8 @@
 namespace OpenFlame\Framework\Cache\Engine;
 use \OpenFlame\Framework\Core;
 
-if(!defined('OpenFlame\\ROOT_PATH')) exit;
-
 /**
- * OpenFlame Web Framework - Cache Engine interface,
+ * OpenFlame Framework - Cache Engine interface,
  * 		Cache engine prototype, declares required methods that a cache engine must define in order to be valid.
  *
  *
@@ -25,9 +24,9 @@ if(!defined('OpenFlame\\ROOT_PATH')) exit;
 interface EngineInterface
 {
 	public function getEngineName();
-	public function build($data, $ttl);
 	public function load($key);
 	public function exists($key);
 	public function destroy($key);
-	public function store($key, $data);
+	public function store($key, $data, $ttl);
+	public function gc(\OpenFlame\Framework\Event\Instance $event = NULL);
 }