+++ Event/Dispatcher.php
@@ -1,8 +1,9 @@
 <?php
 /**
  *
- * @package     OpenFlame Web Framework
- * @copyright   (c) 2010 OpenFlameCMS.com
+ * @package     openflame-framework
+ * @subpackage  event
+ * @copyright   (c) 2010 - 2011 openflame-project.org
  * @license     http://opensource.org/licenses/mit-license.php The MIT License
  * @link        https://github.com/OpenFlame/OpenFlame-Framework
  *
@@ -12,10 +13,8 @@
 namespace OpenFlame\Framework\Event;
 use OpenFlame\Framework\Core;
 
-if(!defined('OpenFlame\\ROOT_PATH')) exit;
-
 /**
- * OpenFlame Web Framework - Event dispatcher object
+ * OpenFlame Framework - Event dispatcher object
  * 	     Provides event dispatcher functionality for ease of extensibility.
  *
  *
@@ -29,6 +28,11 @@ class Dispatcher
 	 */
 	protected $listeners = array();
 
+	/**
+	 * @var array - Array of "unsorted" listeners
+	 */
+	protected $unsorted = array();
+
 	/**#@+
 	 * @var integer - Constants representing the type of listener being interacted with.
 	 */
@@ -38,18 +42,26 @@ class Dispatcher
 	const LISTENER_CALL_USER_FUNC = 4;
 	/**#@-*/
 
+	/**#@+
+	 * @var integer - Constants representing the type of trigger mechanism to use.
+	 */
+	const TRIGGER_NOBREAK = 1;
+	const TRIGGER_MANUALBREAK = 2;
+	const TRIGGER_RETURNBREAK = 3;
+	/**#@-*/
+
 	/**
 	 * Register a new listener with the dispatcher
-	 * @param string $event_type - The type of event type to attach the listener to.
+	 * @param string $event_name - The name of the event to attach the listener to.
 	 * @param integer $priority - The priority for the listener to be registered as, similar to *nix "nice" values for CPU processes (-20 top priority, 20 bottom priority)
 	 * @param callable $listener - The callable reference for the listener.
 	 * @return \OpenFlame\Framework\Event\Dispatcher - Provides a fluent interface.
 	 */
-	public function register($event_type, $priority, $listener)
+	public function register($event_name, $priority, $listener)
 	{
-		if(!isset($this->listeners[$event_type]) || !is_array($this->listeners[$event_type]))
+		if(!isset($this->listeners[$event_name]) || !is_array($this->listeners[$event_name]))
 		{
-			$this->listeners[$event_type] = array();
+			$this->listeners[$event_name] = array();
 		}
 
 		// Handle priority settings (a la UNIX nice values)
@@ -68,27 +80,30 @@ class Dispatcher
 		if($listener instanceof \Closure)
 		{
 			// It's a closure!  <3
-			$listener_type = static::LISTENER_CLOSURE;
+			$listener_type = self::LISTENER_CLOSURE;
 		}
 		elseif(function_exists($listener))
 		{
-			$listener_type = static::LISTENER_FUNCTION;
+			$listener_type = self::LISTENER_FUNCTION;
 		}
 		elseif(is_string($listener) && sizeof(explode('::', $listener, 2)) > 1) // checking to see if we're actually using a static call and doing so properly
 		{
 			$listener = explode('::', $listener, 2);
-			$listener_type = static::LISTENER_STATIC_METHOD;
+			$listener_type = self::LISTENER_STATIC_METHOD;
 		}
 		else
 		{
 			// Worst case scenario.  We HAVE to use call_user_func() now.
-			$listener_type = static::LISTENER_CALL_USER_FUNC;
+			$listener_type = self::LISTENER_CALL_USER_FUNC;
 		}
 
-		$this->listeners[$event_type][$priority][] = array(
+		$this->listeners[$event_name][$priority][] = array(
 			'listener'		=> $listener,
 			'type'			=> $listener_type,
 		);
+
+		// Flag this event as needing a sort before the next event dispatch
+		$this->unsorted[$event_name] = true;
 
 		return $this;
 	}
@@ -106,53 +121,64 @@ class Dispatcher
 	/**
 	 * Dispatch an event to registered listeners
 	 * @param \OpenFlame\Framework\Event\Instance $event - The event to dispatch.
+	 * @param integer $dispatch_type - The type of dispatch method to use (run all listeners, allow listeners to trigger break, break on a non-NULL return value)
 	 * @return \OpenFlame\Framework\Event\Instance - The event dispatched.
 	 */
-	public function trigger(\OpenFlame\Framework\Event\Instance $event)
+	public function trigger(\OpenFlame\Framework\Event\Instance $event, $dispatch_type = self::TRIGGER_NOBREAK)
 	{
-		if(!$this->hasListeners($event->getName()))
+		$event_name = $event->getName();
+
+		// Check to see if this event has ANY listeners - if it doesn't, just bail out.
+		if(!$this->hasListeners($event_name))
 		{
 			return $event;
 		}
 
-		// Ensure the listener priorities are in order
-		ksort($this->listeners[$event->getName()]);
-		foreach($this->listeners[$event->getName()] as $priority => $priority_thread)
+		// On-the-fly priority sorting
+		if(isset($this->unsorted[$event_name]))
 		{
-			for($i = 0, $size = sizeof($priority_thread); $i <= $size - 1; $i++)
+			ksort($this->listeners[$event_name]);
+			unset($this->unsorted[$event_name]);
+		}
+
+		foreach($this->listeners[$event_name] as $priority => $priority_thread)
+		{
+			foreach($priority_thread as $listener_entry)
 			{
-				$listener = $priority_thread[$i]['listener'];
-				$listener_type = $priority_thread[$i]['type'];
+				$listener = $listener_entry['listener'];
+				$listener_type = $listener_entry['type'];
 
-				try
+				// Use faster, quicker methods than call_user_func() for triggering listeners if they're available
+				switch($listener_type)
 				{
-					// Use faster, quicker methods than call_user_func() for triggering listeners if they're available
-					switch($listener_type)
+					case self::LISTENER_CLOSURE:
+					case self::LISTENER_FUNCTION:
+						$return = $listener($event);
+					break;
+
+					case self::LISTENER_STATIC_METHOD:
+						list($class, $method) = $listener;
+						$return = $class::$method($event);
+					break;
+
+					case self::LISTENER_CALL_USER_FUNC:
+					default:
+						$return = call_user_func($listener, $event);
+					break;
+				}
+
+				if($return !== NULL)
+				{
+					$event->setReturn($return);
+					if($dispatch_type = self::TRIGGER_RETURNBREAK)
 					{
-						case static::LISTENER_CLOSURE:
-						case static::LISTENER_FUNCTION:
-							$return = $listener($event);
-							break;
-
-						case static::LISTENER_STATIC_METHOD:
-							list($class, $method) = $listener;
-							$return = $class::$method($event);
-							break;
-
-						case static::LISTENER_CALL_USER_FUNC:
-						default:
-							$return = call_user_func($listener, $event);
-							break;
-					}
-
-					if($return !== NULL)
-					{
-						$event->setReturn($return);
+						return $event; // PHP 5.4 compat -- cannot use "break (int)" anymore, so we just return the $event
 					}
 				}
-				catch(\Exception $e)
+
+				if($dispatch_type = self::TRIGGER_MANUALBREAK && $event->wasBreakTriggered())
 				{
-					throw new \RuntimeException(sprintf('Exception encountered in event listener assigned to event "%1$s"', $event->getName()), 0, $e);
+					return $event; // PHP 5.4 compat -- cannot use "break (int)" anymore, so we just return the $event
 				}
 			}
 		}
@@ -161,64 +187,28 @@ class Dispatcher
 	}
 
 	/**
-	 * Dispatch an event to registered listeners, and checking to see if a listener wants to abort a
+	 * Dispatch an event to registered listeners, and checking to see if a listener wants to abort (and if so, break)
 	 * @param \OpenFlame\Framework\Event\Instance $event - The event to dispatch.
 	 * @return \OpenFlame\Framework\Event\Instance - The event dispatched.
+	 *
+	 * @deprecated since 1.2.0
 	 */
 	public function triggerUntilBreak(\OpenFlame\Framework\Event\Instance $event)
 	{
-		if(!$this->hasListeners($event->getName()))
-		{
-			return $event;
-		}
+		trigger_error('\\OpenFlame\\Framework\\Event\\Dispatcher->triggerUntilBreak() is deprecated', E_USER_DEPRECATED);
+		return $this->trigger($event, self::TRIGGER_MANUALBREAK);
+	}
 
-		// Ensure the listener priorities are in order
-		ksort($this->listeners[$event->getName()]);
-		foreach($this->listeners[$event->getName()] as $priority => $priority_thread)
-		{
-			for($i = 0, $size = sizeof($priority_thread); $i <= $size - 1; $i++)
-			{
-				$listener = $priority_thread[$i]['listener'];
-				$listener_type = $priority_thread[$i]['type'];
-
-				try
-				{
-					// Use faster, quicker methods than call_user_func() for triggering listeners if they're available
-					switch($listener_type)
-					{
-						case static::LISTENER_CLOSURE:
-						case static::LISTENER_FUNCTION:
-							$return = $listener($event);
-							break;
-
-						case static::LISTENER_STATIC_METHOD:
-							list($class, $method) = $listener;
-							$return = $class::$method($event);
-							break;
-
-						case static::LISTENER_CALL_USER_FUNC:
-						default:
-							$return = call_user_func($listener, $event);
-							break;
-					}
-
-					if($return !== NULL)
-					{
-						$event->setReturn($return);
-					}
-				}
-				catch(\Exception $e)
-				{
-					throw new \RuntimeException(sprintf('Exception encountered in event listener assigned to event "%1$s"', $event->getName()), 0, $e);
-				}
-
-				if($event->wasBreakTriggered())
-				{
-					break 2; // break 2 so that we completely break out
-				}
-			}
-		}
-
-		return $event;
+	/**
+	 * Dispatch an event to registered listeners, and checking to see if a listener returned a value yet or not (and if so, break)
+	 * @param \OpenFlame\Framework\Event\Instance $event - The event to dispatch.
+	 * @return \OpenFlame\Framework\Event\Instance - The event dispatched.
+	 *
+	 * @deprecated since 1.2.0
+	 */
+	public function triggerUntilReturn(\OpenFlame\Framework\Event\Instance $event)
+	{
+		trigger_error('\\OpenFlame\\Framework\\Event\\Dispatcher->triggerUntilReturn() is deprecated', E_USER_DEPRECATED);
+		return $this->trigger($event, self::TRIGGER_RETURNBREAK);
 	}
 }