+++ URL/BuilderProxy.php
@@ -1,8 +1,9 @@
 <?php
 /**
  *
- * @package     OpenFlame Web Framework
- * @copyright   (c) 2010 OpenFlameCMS.com
+ * @package     openflame-framework
+ * @subpackage  url
+ * @copyright   (c) 2010 - 2011 openflame-project.org
  * @license     http://opensource.org/licenses/mit-license.php The MIT License
  * @link        https://github.com/OpenFlame/OpenFlame-Framework
  *
@@ -12,10 +13,8 @@
 namespace OpenFlame\Framework\URL;
 use OpenFlame\Framework\Core;
 
-if(!defined('OpenFlame\\ROOT_PATH')) exit;
-
 /**
- * OpenFlame Web Framework - Template proxy object for internal URL construction,
+ * OpenFlame Framework - Template proxy object for internal URL construction,
  * 	     Provides a near-seamless method of generating internal URLs inside of templates.
  *
  *
@@ -59,7 +58,7 @@ class BuilderProxy
 			return '';
 		}
 
-		$url = $this->builder->getBaseURL() . '/' . vsprintf($pattern, $arguments);
+		$url = rtrim($this->builder->getBaseURL(), '/') . '/' . vsprintf($pattern, $arguments);
 
 		if($this->extra_get_data === false)
 		{